requirements.txt
anyio==3.6.1
asgiref==3.4.1
beanie==1.7.2
click==7.1.2
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.83.0
fastapi-login==1.8.2
h11==0.13.0
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
Jinja2==3.0.3
MarkupSafe==2.0.1
motor==2.5.1
multidict==5.2.0
passlib==1.7.4
pydantic==1.9.2
PyJWT==2.4.0
pymongo==3.12.3
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.16.0
yarl==1.7.2
zipp==3.6.0
